requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.1.0
azure-core==1.35.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.0
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b42
azure-storage-blob==12.19.0
bcrypt==4.1.2
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
Django==5.1.5
django-apscheduler==0.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-redis==6.0.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dotenv==0.9.9
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.18.0
fixedint==0.1.6
frozenlist==1.5.0
fsspec==2025.3.2
gunicorn==23.0.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.1.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-urllib==0.58b0
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.52
propcache==0.3.0
psutil==7.1.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
regex==2024.11.6
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==2.0.0
safetensors==0.5.3
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
twilio==9.4.6
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.11.0
wrapt==1.17.3
yarl==1.18.3
zipp==3.23.0
